requirements.txt
streamlit
pandas
scikit-learn
plotly
groq
python-dotenv
